--- module/decomposition/assets.py
@@ -1,14 +1,0 @@
-from module.base.button import Button
-from module.base.template import Template
-
-# This file was automatically generated by dev_tools.button_extract.
-# Don't modify it manually.
-
-BOX_USE10_1 = Button(area={'cn': (683, 477, 855, 532), 'en': (683, 477, 855, 532), 'jp': (683, 477, 855, 532)}, color={'cn': (108, 153, 210), 'en': (108, 153, 210), 'jp': (108, 153, 210)}, button={'cn': (683, 477, 855, 532), 'en': (683, 477, 855, 532), 'jp': (683, 477, 855, 532)}, file={'cn': './assets/cn/decomposition/BOX_USE10_1.png', 'en': './assets/cn/decomposition/BOX_USE10_1.png', 'jp': './assets/cn/decomposition/BOX_USE10_1.png'})
-BOX_USE10_2 = Button(area={'cn': (748, 477, 920, 532), 'en': (748, 477, 920, 532), 'jp': (748, 477, 920, 532)}, color={'cn': (107, 153, 209), 'en': (107, 153, 209), 'jp': (107, 153, 209)}, button={'cn': (748, 477, 920, 532), 'en': (748, 477, 920, 532), 'jp': (748, 477, 920, 532)}, file={'cn': './assets/cn/decomposition/BOX_USE10_2.png', 'en': './assets/cn/decomposition/BOX_USE10_2.png', 'jp': './assets/cn/decomposition/BOX_USE10_2.png'})
-INFORM_BOX = Button(area={'cn': (340, 198, 941, 261), 'en': (340, 198, 941, 261), 'jp': (340, 198, 941, 261)}, color={'cn': (114, 125, 169), 'en': (114, 125, 169), 'jp': (114, 125, 169)}, button={'cn': (340, 198, 941, 261), 'en': (340, 198, 941, 261), 'jp': (340, 198, 941, 261)}, file={'cn': './assets/cn/decomposition/INFORM_BOX.png', 'en': './assets/cn/decomposition/INFORM_BOX.png', 'jp': './assets/cn/decomposition/INFORM_BOX.png'})
-INFORM_FULL = Button(area={'cn': (346, 169, 933, 230), 'en': (346, 169, 933, 230), 'jp': (346, 169, 933, 230)}, color={'cn': (110, 122, 166), 'en': (110, 122, 166), 'jp': (110, 122, 166)}, button={'cn': (346, 169, 933, 230), 'en': (346, 169, 933, 230), 'jp': (346, 169, 933, 230)}, file={'cn': './assets/cn/decomposition/INFORM_FULL.png', 'en': './assets/cn/decomposition/INFORM_FULL.png', 'jp': './assets/cn/decomposition/INFORM_FULL.png'})
-MATERIAL = Button(area={'cn': (1105, 651, 1259, 703), 'en': (1105, 651, 1259, 703), 'jp': (1105, 651, 1259, 703)}, color={'cn': (169, 133, 78), 'en': (169, 133, 78), 'jp': (169, 133, 78)}, button={'cn': (1105, 651, 1259, 703), 'en': (1105, 651, 1259, 703), 'jp': (1105, 651, 1259, 703)}, file={'cn': './assets/cn/decomposition/MATERIAL.png', 'en': './assets/cn/decomposition/MATERIAL.png', 'jp': './assets/cn/decomposition/MATERIAL.png'})
-STORAGE_AREA = Button(area={'cn': (126, 57, 1232, 650), 'en': (126, 57, 1232, 650), 'jp': (126, 57, 1232, 650)}, color={'cn': (255, 255, 255), 'en': (255, 255, 255), 'jp': (255, 255, 255)}, button={'cn': (126, 57, 1232, 650), 'en': (126, 57, 1232, 650), 'jp': (126, 57, 1232, 650)}, file={'cn': './assets/cn/decomposition/STORAGE_AREA.png', 'en': './assets/cn/decomposition/STORAGE_AREA.png', 'jp': './assets/cn/decomposition/STORAGE_AREA.png'})
-STORAGE_CHECK = Button(area={'cn': (1225, 16, 1257, 45), 'en': (1225, 16, 1257, 45), 'jp': (1225, 16, 1257, 45)}, color={'cn': (108, 124, 151), 'en': (108, 124, 151), 'jp': (108, 124, 151)}, button={'cn': (1225, 16, 1257, 45), 'en': (1225, 16, 1257, 45), 'jp': (1225, 16, 1257, 45)}, file={'cn': './assets/cn/decomposition/STORAGE_CHECK.png', 'en': './assets/cn/decomposition/STORAGE_CHECK.png', 'jp': './assets/cn/decomposition/STORAGE_CHECK.png'})
-STORAGE_OPEN = Button(area={'cn': (328, 667, 484, 713), 'en': (328, 667, 484, 713), 'jp': (328, 667, 484, 713)}, color={'cn': (93, 111, 143), 'en': (93, 111, 143), 'jp': (93, 111, 143)}, button={'cn': (328, 667, 484, 713), 'en': (328, 667, 484, 713), 'jp': (328, 667, 484, 713)}, file={'cn': './assets/cn/decomposition/STORAGE_OPEN.png', 'en': './assets/cn/decomposition/STORAGE_OPEN.png', 'jp': './assets/cn/decomposition/STORAGE_OPEN.png'})
